refactor(ds): update Avatar components - extract utils - index file export
