+++ ui/design-system-core/src/components/Avatar/index.tsx
@@ -5,8 +5,11 @@ import { Profile } from '@akashaorg/typings/ui';
 import Box from '../Box';
 
 import AvatarImage from './avatar-image';
-
-import { getAvatarFromSeed } from '../../utils/get-avatar-from-seed';
+import {
+  generateActiveOverlayClass,
+  generateAvatarContainerStyle,
+  getAvatarFromSeed,
+} from '../../utils';
 
 export type AvatarSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
 
@@ -28,30 +31,6 @@ export type AvatarProps = {
   isClickable?: boolean;
   customStyle?: string;
   onClick?: React.MouseEventHandler<HTMLDivElement>;
-};
-
-export const avatarSizesMap = {
-  xs: 'w-6 h-6',
-  sm: 'w-8 h-8',
-  md: 'w-10 h-10',
-  lg: 'w-16 h-16',
-  xl: 'w-20 h-20',
-  xxl: 'w-[7.5rem] h-[7.5rem]',
-};
-
-export const avatarBorderSizesMap = {
-  xs: '1',
-  sm: '2',
-  md: '4',
-  lg: '8',
-  xl: '[12px]',
-  xxl: '[16px]',
-};
-
-export const avatarBorderColorsMap = {
-  white: 'white',
-  darkerBlue: 'grey2',
-  accent: 'secondaryDark',
 };
 
 const Avatar: React.FC<AvatarProps> = props => {
@@ -78,16 +57,16 @@ const Avatar: React.FC<AvatarProps> = props => {
     avatarImageFallback = `${publicImgPath}/avatar-placeholder-${seed}.webp`;
   }
 
-  const containerStyle = `box-border cursor-${
-    isClickable ? 'pointer' : 'default'
-  } select-none relative overflow-hidden ${avatarSizesMap[size]} rounded-full bg-${
-    backgroundColor ? backgroundColor : 'white'
-  } border-${border ? avatarBorderSizesMap[border] : '0'} border-${
-    borderColor ? avatarBorderColorsMap[borderColor] : 'transparent'
-  } ${customStyle}`;
+  const containerStyle = generateAvatarContainerStyle({
+    size,
+    border,
+    borderColor,
+    customStyle,
+    isClickable,
+    backgroundColor,
+  });
 
-  const activeOverlayClass =
-    'bg-grey6 dark:bg-grey6 opacity-25 z-10 absolute top-0 left-0 w-full h-full';
+  const activeOverlayClass = generateActiveOverlayClass();
 
   return (
     <Box customStyle={containerStyle} onClick={onClick}>
